+++ src/ebm/samplers/__init__.py
@@ -1,6 +1,6 @@
 from ebm.samplers.annealed import AnnealedLangevinDynamics
 from ebm.samplers.base import Sampler
-from ebm.samplers.discrete import GibbsWithGradients
+from ebm.samplers.discrete import CategoricalGibbsWithGradients, GibbsWithGradients
 from ebm.samplers.hmc import HMC
 from ebm.samplers.langevin import MALA, LangevinDynamics
 
@@ -10,5 +10,6 @@ __all__ = [
     "MALA",
     "HMC",
     "GibbsWithGradients",
+    "CategoricalGibbsWithGradients",
     "AnnealedLangevinDynamics",
 ]